169 - Rename suite's components
