+++ frontend/src/components/suite/LogMessagesTable.styles.js
